refactor: 프로그래스바 컴포넌트 위치 바꿈 #165
